Restructure models.
