+++ functions/src/models/categories.ts
@@ -1,0 +1,11 @@
+import { QuizRound } from './questions';
+
+interface Categories {
+  [index: string]: QuizRound;
+}
+
+enum Category {
+  SPORT = 'sport',
+}
+
+export { Category, Categories };
